Make gator root command
Per decision in weekly meeting, make the "gator" root command and "test"
the subcommand in charge of running the tests. This frees us up to add
other subcommands that we want to do later.
